# Conflicts:
#	jira/handlers/projects.handler.go
#	jira/models/projects.model.go

+++ jira/models/issues.model.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"io"
 	. "jira/common/db"
+	"strconv"
 
 	// . "jira/common/helpers"
 
@@ -160,14 +161,36 @@ func (pm *IssuesModel) GetById(id string) ([]Issue, error) {
 	return ListIssues, nil
 }
 
-func (pm *IssuesModel) Create(r io.ReadCloser) ([]Issue, error) {
+func (pm *IssuesModel) Create(r io.ReadCloser, userId int64) ([]Issue, error) {
 	var issue Issue
 	json.NewDecoder(r).Decode(&issue)
-	query := fmt.Sprintf(
+	query := fmt.Sprintf("select C.permission_id from new_jira_user_project_role A, new_jira_role_permission B, new_jira_permission C where A.user_id = '%v' and A.project_key = '%v' and A.role_id = B.role_id and B.permission_id = C.permission_id", userId, issue.Project)
+	rows, err := DbOracle.Db.Query(query)
+	if err != nil {
+		return nil, err
+	}
+	var listPermission_id []int
+	for rows.Next() {
+		var permission_id int
+		rows.Scan(&permission_id)
+		listPermission_id = append(listPermission_id, permission_id)
+	}
+	if len(listPermission_id) == 0 {
+		return nil, nil
+	}
+	for i, v := range listPermission_id {
+		if v == 5 {
+			break
+		}
+		if i == len(listPermission_id)-1 {
+			return nil, nil
+		}
+	}
+	query = fmt.Sprintf(
 		`INSERT INTO NEW_JIRA_ISSUE (KEY,NAME,PROJECT,ISSUE_TYPE,ID,ICON)
 		VALUES ('%v', '%v', '%v', '%v', SEQ_NEW_JIRA_ISSUE.nextval, '%v')`,
 		issue.Key, issue.Name, issue.Project, issue.Issue_Type, issue.Icon)
-	_, err := DbOracle.Db.Exec(query)
+	_, err = DbOracle.Db.Exec(query)
 	if err == nil {
 		_, err := FieldsModels.Create(issue.Fields)
 		if err != nil {
@@ -182,7 +205,11 @@ func (pm *IssuesModel) Create(r io.ReadCloser) ([]Issue, error) {
 				rowsLastRecord.Scan(&issue.Key, &issue.Name, &issue.Project, &issue.Issue_Type, &issue.Id, &issue.Icon)
 				ListIssues = append(ListIssues, issue)
 			}
-			return ListIssues, nil
+			issues, err := IssuesModels.GetById(strconv.Itoa(ListIssues[0].Id))
+			if err != nil {
+				return nil, err
+			}
+			return issues, nil
 		} else {
 			return nil, err
 		}
@@ -191,11 +218,33 @@ func (pm *IssuesModel) Create(r io.ReadCloser) ([]Issue, error) {
 	}
 }
 
-func (pm *IssuesModel) Update(r io.ReadCloser, id string) ([]Issue, error) {
+func (pm *IssuesModel) Update(r io.ReadCloser, id string, userId int64) ([]Issue, error) {
 	var issue Issue
 	json.NewDecoder(r).Decode(&issue)
-	query := fmt.Sprintf(`UPDATE new_jira_issue SET name = '%v' WHERE id = '%v'`, issue.Name, id)
-	_, err := DbOracle.Db.Exec(query)
+	query := fmt.Sprintf("select C.permission_id from new_jira_user_project_role A, new_jira_role_permission B, new_jira_permission C where A.user_id = '%v' and A.project_key = '%v' and A.role_id = B.role_id and B.permission_id = C.permission_id", userId, issue.Project)
+	rows, err := DbOracle.Db.Query(query)
+	if err != nil {
+		return nil, err
+	}
+	var listPermission_id []int
+	for rows.Next() {
+		var permission_id int
+		rows.Scan(&permission_id)
+		listPermission_id = append(listPermission_id, permission_id)
+	}
+	if len(listPermission_id) == 0 {
+		return nil, nil
+	}
+	for i, v := range listPermission_id {
+		if v == 7 {
+			break
+		}
+		if i == len(listPermission_id)-1 {
+			return nil, nil
+		}
+	}
+	query = fmt.Sprintf(`UPDATE new_jira_issue SET name = '%v' WHERE id = '%v'`, issue.Name, id)
+	_, err = DbOracle.Db.Exec(query)
 	if err == nil {
 		query := fmt.Sprintf("delete from new_jira_field where issue = '%v'", issue.Key)
 		_, err = DbOracle.Db.Exec(query)
@@ -206,18 +255,44 @@ func (pm *IssuesModel) Update(r io.ReadCloser, id string) ([]Issue, error) {
 		if err != nil {
 			return nil, err
 		}
-		return nil, nil
+		issues, err := IssuesModels.GetById(id)
+		if err != nil {
+			return nil, err
+		}
+		return issues, nil
 	} else {
 		return nil, err
 	}
 }
 
-func (pm *IssuesModel) Delete(id string) ([]Issue, error) {
+func (pm *IssuesModel) Delete(id string, userId int64) ([]Issue, error) {
 	issues, err := IssuesModels.GetById(id)
 	if err != nil {
 		return nil, err
 	}
-	query := fmt.Sprintf("delete from new_jira_field where issue = '%v'", issues[0].Key)
+	query := fmt.Sprintf("select C.permission_id from new_jira_user_project_role A, new_jira_role_permission B, new_jira_permission C where A.user_id = '%v' and A.project_key = '%v' and A.role_id = B.role_id and B.permission_id = C.permission_id", userId, issues[0].Project)
+	rows, err := DbOracle.Db.Query(query)
+	if err != nil {
+		return nil, err
+	}
+	var listPermission_id []int
+	for rows.Next() {
+		var permission_id int
+		rows.Scan(&permission_id)
+		listPermission_id = append(listPermission_id, permission_id)
+	}
+	if len(listPermission_id) == 0 {
+		return nil, nil
+	}
+	for i, v := range listPermission_id {
+		if v == 6 {
+			break
+		}
+		if i == len(listPermission_id)-1 {
+			return nil, nil
+		}
+	}
+	query = fmt.Sprintf("delete from new_jira_field where issue = '%v'", issues[0].Key)
 	_, err = DbOracle.Db.Exec(query)
 	if err != nil {
 		return nil, err
